# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SoulForge API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /goals/{goalId}/subtasks:
    get:
      operationId: listSubtasks
      tags: [subtasks]
      summary: List subtasks for a goal
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of subtasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subtask"
    post:
      operationId: createSubtask
      tags: [subtasks]
      summary: Create a subtask
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubtaskInput"
      responses:
        "201":
          description: Created subtask
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subtask"
  /goals/{goalId}/subtasks/{subtaskId}:
    patch:
      operationId: updateSubtask
      tags: [subtasks]
      summary: Update a subtask
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: subtaskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubtaskUpdate"
      responses:
        "200":
          description: Updated subtask
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subtask"
    delete:
      operationId: deleteSubtask
      tags: [subtasks]
      summary: Delete a subtask
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: subtaskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /goals/{goalId}/subtasks/{subtaskId}/complete:
    post:
      operationId: completeSubtask
      tags: [subtasks]
      summary: Complete a subtask (awards XP, damages boss)
      parameters:
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: subtaskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: XP and boss damage result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/XPResult"

  # HABITS
...
  /skills/{skillId}/equip:
    post:
      operationId: equipSkill
      tags: [skills]
      summary: Equip a skill
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    delete:
      operationId: unequipSkill
      tags: [skills]
      summary: Unequip a skill
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"

  # CALENDAR
...
  /admin/skills:
    get:
      operationId: adminListSkills
      tags: [admin]
      summary: List all skill definitions
      responses:
        "200":
          description: Skills list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SkillDefinition"
    post:
      operationId: adminCreateSkill
      tags: [admin]
      summary: Create a skill definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillDefinitionInput"
      responses:
        "201":
          description: Created skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillDefinition"
  /admin/skills/{skillId}:
    patch:
      operationId: adminUpdateSkill
      tags: [admin]
      summary: Update a skill definition
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SkillDefinitionInput"
      responses:
        "200":
          description: Updated skill
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillDefinition"
    delete:
      operationId: adminDeleteSkill
      tags: [admin]
      summary: Delete a skill definition
      parameters:
        - name: skillId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # AUTH
...
    Player:
      type: object
      required: [id, userId, username, level, xp, xpToNextLevel, hp, maxHp, gold, strength, endurance, dexterity, faith, avatarClass, equippedSkills, createdAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        username:
          type: string
        level:
          type: integer
        xp:
          type: integer
        xpToNextLevel:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        gold:
          type: integer
        strength:
          type: integer
        endurance:
          type: integer
        dexterity:
          type: integer
        faith:
          type: integer
        avatarClass:
          type: string
        equippedSkills:
          type: array
          items:
            type: string
        createdAt:
          type: string
...
    XPResult:
      type: object
      required: [xpGained, player, leveledUp, loot]
      properties:
        xpGained:
          type: integer
        player:
          $ref: "#/components/schemas/Player"
        leveledUp:
          type: boolean
        bossHpReduced:
          type: ["integer", "null"]
        loot:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"

    # GOALS
...
    Subtask:
      type: object
      required: [id, goalId, userId, title, completed, xpReward, createdAt, completedAt]
      properties:
        id:
          type: string
          format: uuid
        goalId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        title:
          type: string
        completed:
          type: boolean
        xpReward:
          type: integer
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
    SubtaskInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        xpReward:
          type: integer
    SubtaskUpdate:
      type: object
      properties:
        title:
          type: string
        completed:
          type: boolean

    # HABITS
...
    InventoryItem:
      type: object
      required: [id, userId, name, description, rarity, type, obtainedAt]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        rarity:
          type: string
          enum: [common, uncommon, rare, epic, legendary]
        type:
          type: string
          enum: [weapon, armor, accessory, consumable, trophy]
        obtainedAt:
          type: string

    # SKILLS
...
    SkillDefinition:
      type: object
      required: [id, name, description, effect, levelRequired]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        effect:
          type: string
        levelRequired:
          type: integer
    SkillDefinitionInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        effect:
          type: string
        levelRequired:
          type: integer
...